backend/requirements.txt
Flask==3.0.0
paho-mqtt==1.6.1

